Initial BNE Implementation
Baseline BNE implementation. There is a possibility I got something
wrong here, but I basically just copied over everything I saw from the
existing BEQ implementation and made the required modifications. I don't
THINK there is anything special about the BNE implementation that'd
require me to modify anything anywhere that doesn't already have BEQ
code for it, but if I'm wrong and there is then this implementation is
missing whatever that component is. (though again, I'm pretty sure this
is a rather simple conversion)

One important note is that I noticed the initial definitions seem to be
a bit inconsistent. In lazyvim some of the spaces have greyed out ">"
between them and some are just normal spaces. I think the difference is
tab-spacing vs normal-space-spacing and it might not matter, but it's
worth noting the descrepancy. In either case, just to be safe I copied
the BEQ definition line exactly and just modified it, so if the
invisible ">" symbols do have some semantic meaning my BNE definition
should still be valid since it's defined in the same way as BEQ was.